ut_course_search_back/requirements.txt
beautifulsoup4==4.12.2
Flask==3.0.0
Flask-Cors==4.0.0
flatbuffers==23.5.8
huggingface-hub==0.17.3
langcodes==3.3.0
llama-index==0.9.13
openai==1.3.7
pandas==2.1.2
pinecone-client==2.2.4
python-dotenv==1.0.0
trafilatura==1.6.0
transformers==4.34.1
gunicorn==21.2.0
torch==2.1.0
faiss-cpu==1.7.4
